Scales

+++ app/src/main/java/com/bsw/snakes/entities/SnakePoints.java
@@ -1,27 +1,27 @@
 package com.bsw.snakes.entities;
 
 public class SnakePoints {
-    int xPosition;
-    int yPosition;
+    float xPosition;
+    float yPosition;
 
-    public SnakePoints(int xPosition, int yPosition) {
+    public SnakePoints(float xPosition, float yPosition) {
         this.xPosition = xPosition;
         this.yPosition = yPosition;
     }
 
-    public int getxPosition() {
+    public float getxPosition() {
         return xPosition;
     }
 
-    public void setxPosition(int xPosition) {
+    public void setxPosition(float xPosition) {
         this.xPosition = xPosition;
     }
 
-    public int getyPosition() {
+    public float getyPosition() {
         return yPosition;
     }
 
-    public void setyPosition(int yPosition) {
+    public void setyPosition(float yPosition) {
         this.yPosition = yPosition;
     }
 }
